+++ Cargo.toml
@@ -73,6 +73,7 @@ members = [
     "language/move-prover/spec-lang",
     "language/move-prover/test-utils",
     "language/move-prover/stackless-bytecode-generator",
+    "language/move-vm/natives",
     "language/move-vm/runtime",
     "language/move-vm/state",
     "language/move-vm/types",
