+++ language/move-vm/natives/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "move-vm-runtime"
+name = "move-vm-natives"
 version = "0.1.0"
 authors = ["Libra Association <opensource@libra.org>"]
-description = "Core Move execution logic"
+description = "Move VM native functions"
 repository = "https://github.com/libra/libra"
 homepage = "https://libra.org"
 license = "Apache-2.0"
@@ -12,29 +12,18 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bit-vec = "0.6.1"
 once_cell = "1.3.1"
 mirai-annotations = "1.7.0"
+sha2 = "0.8.0"
 
-bytecode-verifier = { path = "../../bytecode-verifier", version = "0.1.0" }
 libra-crypto = { path = "../../../crypto/crypto", version = "0.1.0" }
 libra-logger = { path = "../../../common/logger", version = "0.1.0" }
 libra-types = { path = "../../../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../../../common/workspace-hack", version = "0.1.0" }
 move-core-types = { path = "../../move-core/types", version = "0.1.0" }
-move-vm-natives = { path = "../natives", version = "0.1.0" }
 move-vm-types = { path = "../types", version = "0.1.0" }
 vm = { path = "../../vm", version = "0.1.0" }
-
-[dev-dependencies]
-anyhow = "1.0"
-hex = "0.4.2"
-proptest = "0.9.6"
-
-move-vm-types = { path = "../types", version = "0.1.0", features = ["fuzzing"] }
-
-compiler = { path = "../../compiler", version = "0.1.0" }
-libra-state-view = { path = "../../../storage/state-view", version = "0.1.0" }
-move-vm-state = { path = "../state", version = "0.1.0" }
 
 [features]
 default = []
