+++ language/move-vm/natives/src/event.rs
@@ -1,17 +1,17 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{
-    loaded_data::runtime_types::Type,
-    native_functions::{context::NativeContext, dispatch::NativeResult},
-    values::Value,
-};
 use libra_types::{
     contract_event::ContractEvent,
     event::EventKey,
     vm_error::{StatusCode, VMStatus},
 };
 use move_core_types::gas_schedule::ZERO_GAS_UNITS;
+use move_vm_types::{
+    loaded_data::runtime_types::Type,
+    natives::function::{NativeContext, NativeResult},
+    values::Value,
+};
 use std::{collections::VecDeque, convert::TryFrom};
 use vm::errors::VMResult;
 
@@ -20,14 +20,8 @@ pub fn native_emit_event(
     ty_args: Vec<Type>,
     mut arguments: VecDeque<Value>,
 ) -> VMResult<NativeResult> {
-    if ty_args.len() != 1 {
-        return Err(
-            VMStatus::new(StatusCode::VERIFIER_INVARIANT_VIOLATION).with_message(format!(
-                "write_to_event_storage expects 1 type argument got {}.",
-                ty_args.len()
-            )),
-        );
-    }
+    debug_assert!(ty_args.len() == 1);
+    debug_assert!(arguments.len() == 3);
 
     let mut ty_args = context.convert_to_fat_types(ty_args)?;
     let ty = ty_args.pop().unwrap();
