+++ language/move-vm/natives/src/lib.rs
@@ -1,0 +1,11 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+#[macro_use]
+extern crate move_vm_types;
+
+pub mod account;
+pub mod event;
+pub mod hash;
+pub mod lcs;
+pub mod signature;
