+++ language/move-vm/runtime/src/lib.rs
@@ -16,7 +16,6 @@ mod gas_meter;
 mod interpreter;
 mod loader;
 mod move_vm;
-mod native_context;
 mod native_functions;
 mod runtime;
