--- language/move-vm/types/src/native_functions/dispatch.rs
@@ -1,57 +1,0 @@
-// Copyright (c) The Libra Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-use crate::values::Value;
-use libra_types::vm_error::VMStatus;
-use move_core_types::gas_schedule::{
-    AbstractMemorySize, CostTable, GasAlgebra, GasCarrier, GasUnits, NativeCostIndex,
-};
-use vm::errors::VMResult;
-
-/// Result of a native function execution that requires charges for execution cost.
-///
-/// An execution that causes an invariant violation would not return a `NativeResult` but
-/// return a `VMResult` error directly.
-/// All native functions must return a `VMResult<NativeResult>` where an `Err` is returned
-/// when an error condition is met that should not charge for the execution. A common example
-/// is a VM invariant violation which should have been forbidden by the verifier.
-/// Errors (typically user errors and aborts) that are logically part of the function execution
-/// must be expressed in a `NativeResult` via a cost and a VMStatus.
-pub struct NativeResult {
-    /// The cost for running that function, whether successfully or not.
-    pub cost: GasUnits<GasCarrier>,
-    /// Result of execution. This is either the return values or the error to report.
-    pub result: VMResult<Vec<Value>>,
-}
-
-impl NativeResult {
-    /// Return values of a successful execution.
-    pub fn ok(cost: GasUnits<GasCarrier>, values: Vec<Value>) -> Self {
-        NativeResult {
-            cost,
-            result: Ok(values),
-        }
-    }
-
-    /// `VMStatus` of a failed execution. The failure is a runtime failure in the function
-    /// and not an invariant failure of the VM which would raise a `VMResult` error directly.
-    pub fn err(cost: GasUnits<GasCarrier>, err: VMStatus) -> Self {
-        NativeResult {
-            cost,
-            result: Err(err),
-        }
-    }
-}
-
-pub fn native_gas(table: &CostTable, key: NativeCostIndex, size: usize) -> GasUnits<GasCarrier> {
-    let gas_amt = table.native_cost(key);
-    let memory_size = AbstractMemorySize::new(size as GasCarrier);
-    gas_amt.total().mul(memory_size)
-}
-
-#[macro_export]
-macro_rules! pop_arg {
-    ($arguments:ident, $t:ty) => {{
-        $arguments.pop_back().unwrap().value_as::<$t>()?
-    }};
-}
